// src/enums/claim-codes.ts
export enum ClaimCode {
  DMA = 'DMA',
  DME = 'DME',
  DMV = 'DMV',
  DMD = 'DMD',
  UMA = 'UMA',
  UME = 'UME',
  UMV = 'UMV',
  UMD = 'UMD',
  CMA = 'CMA',
  CME = 'CME',
  CMV = 'CMV',
  CMD = 'CMD',
  CliMgA = 'CliMgA',
  CliMgE = 'CliMgE',
  CliMgV = 'CliMgV',
  CliMgD = 'CliMgD',
  VSCA = 'VSCA',
  VSCE = 'VSCE',
  VSCV = 'VSCV',
  VSCD = 'VSCD',
  JMA = 'JMA',
  JME = 'JME',
  JMV = 'JMV',
  JMD = 'JMD',
  JAMA = 'JAMA',
  JAMV = 'JAMV',
  JAMC = 'JAMC',
  RMA = 'RMA',
  RME = 'RME',
  RMV = 'RMV',
  RMD = 'RMD',
  PMA = 'PMA',
  PME = 'PME',
  PMV = 'PMV',
  PMD = 'PMD',
  RAMA = 'RAMA',
  RAMV = 'RAMV',
  RAMC = 'RAMC',
  OLMV = 'OLMV',
  QRTOA = 'QRTOA',
  QRTOE = 'QRTOE',
  QRTOV = 'QRTOV',
  QRTOD = 'QRTOD',
  TRMA = 'TRMA',
  TRME = 'TRME',
  TRMV = 'TRMV',
  TRMD = 'TRMD',
  DRMA = 'DRMA',
  DRME = 'DRME',
  DRMV = 'DRMV',
  DRMD = 'DRMD',
  AEMA = 'AEMA',
  AEME = 'AEME',
  AEMV = 'AEMV',
  AEMD = 'AEMD',
  MTCMA = 'MTCMA',
  MTCME = 'MTCME',
  MTCMV = 'MTCMV',
  MTCMD = 'MTCMD',
  CSSQA = 'CSSQA',
  CSSQE = 'CSSQE',
  CSSQV = 'CSSQV',
  CSSQD = 'CSSQD',
  MAPNA = 'MAPNA',
  MAPNE = 'MAPNE',
  MAPNV = 'MAPNV',
  MAPND = 'MAPND',
  RV = 'RV',
  UPA = 'UPA',
  UPE = 'UPE',
  UPD = 'UPD',
  UPV = 'UPV',
  OME = 'OME',
  OMV = 'OMV',
  OSMV = 'OSMV',
  ARAA = 'ARAA',
  ARAD = 'ARAD',
  CGMA = 'CGMA',
  CGME = 'CGME',
  CGMV = 'CGMV',
  CGMD = 'CGMD',
  CRV = 'CRV',
}